+++ no-bypass/main.c
@@ -43,7 +43,7 @@ struct sockaddr_in new_sockaddr (const char *ip, uint16_t port)
     return addr;
 }
 
-void start_server (uint32_t iters)
+void start_server (uint64_t iters)
 {
     int socket = new_socket ();
     // wait for the client to connect
@@ -102,7 +102,7 @@ int send_single_packet (char *buf, const int packet_idx, struct sockaddr_ll *add
     return sendto (socket, buf, PACKET_SIZE, 0, (struct sockaddr *) addr, sizeof (*addr));
 }
 
-void start_client (uint32_t iters, uint64_t interval, const char *server_ip)
+void start_client (uint64_t iters, uint64_t interval, const char *server_ip)
 {
     int socket = new_socket ();
     struct sockaddr_in server_addr = new_sockaddr (server_ip, PINGPONG_UDP_PORT);
@@ -136,7 +136,7 @@ void start_client (uint32_t iters, uint64_t interval, const char *server_ip)
 int main (int argc, char **argv)
 {
 #if SERVER
-    uint32_t iters = 0;
+    uint64_t iters = 0;
     if (!nobypass_parse_args (argc, argv, &iters))
     {
         nobypass_print_usage (argv[0]);
@@ -149,7 +149,7 @@ int main (int argc, char **argv)
 #else
     uint32_t persistence_flag = PERSISTENCE_M_ALL_TIMESTAMPS;
 
-    uint32_t iters = 0;
+    uint64_t iters = 0;
     uint64_t interval = 0;
     char *server_ip = NULL;
     if (!nobypass_parse_args (argc, argv, &iters, &interval, &server_ip, &persistence_flag))
@@ -165,7 +165,7 @@ int main (int argc, char **argv)
         return EXIT_FAILURE;
     }
 
-    LOG (stdout, "Starting client with iters=%u, interval=%lu, server_ip=%s\n", iters, interval, server_ip);
+    LOG (stdout, "Starting client with iters=%lu, interval=%lu, server_ip=%s\n", iters, interval, server_ip);
 
     start_client (iters, interval, server_ip);
 #endif
